ingestion_function_py/requirements.txt
azure-functions==1.23.0
azure-storage-blob==12.23.1
azure-ai-documentintelligence==1.0.2
azure-core==1.30.2
azure-identity==1.16.0
azure-search-documents==11.6.0
azure-storage-queue==12.11.0
langchain-core==0.2.24
langchain-community==0.2.10
langchain-openai==0.1.19
langchain-text-splitters==0.2.2
openai==1.37.1
httpx==0.27.0
pandas==2.2.2
openpyxl==3.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2